update log line to search
