+++ Tests/E2E/C2DMessageTest.cs
@@ -7,7 +7,6 @@ namespace LoRaWan.Tests.E2E
     using System.Globalization;
     using System.Linq;
     using System.Threading.Tasks;
-    using LoRaTools;
     using LoRaTools.CommonAPI;
     using LoRaWan.Tests.Common;
     using Xunit;
@@ -575,142 +574,6 @@ namespace LoRaWan.Tests.E2E
             }
 
             Assert.True(foundReceivePacketCount > 0, $"Could not find lora receiving message '{expectedRxSerial}'");
-        }
-
-        /// <summary>
-        /// Ensures that data rate is updated on a device after a LinkADRCmd MAC Command is sent as C2D message.
-        /// </summary>
-        [RetryFact]
-        public async Task Data_Rate_Is_Updated_When_C2D_With_LinkADRCmd_Received()
-        {
-            const int messagesToSend = 10;
-            const int warmUpMessageCount = 2;
-
-            var device = TestFixtureCi.Device32_ABP;
-            LogTestStart(device);
-
-            // Setup LoRa device properties
-            await ArduinoDevice.setDeviceModeAsync(LoRaArduinoSerial._device_mode_t.LWABP);
-            await ArduinoDevice.setIdAsync(device.DevAddr, device.DeviceID, device.AppEUI);
-            await ArduinoDevice.setKeyAsync(device.NwkSKey, device.AppSKey, device.AppKey);
-
-            // Setup protocol properties
-            // Start with DR5
-            await ArduinoDevice.SetupLora(TestFixtureCi.Configuration.LoraRegion, LoRaArduinoSerial._data_rate_t.DR5, 4, true);
-            await TestFixture.CleanupC2DDeviceQueueAsync(device.DeviceID);
-
-            // Sends 2x unconfirmed messages
-            for (var i = 1; i <= warmUpMessageCount; ++i)
-            {
-                var msg = PayloadGenerator.Next().ToString(CultureInfo.InvariantCulture);
-                Log($"{device.DeviceID}: Sending unconfirmed '{msg}' {i}/{messagesToSend}");
-                await ArduinoDevice.transferPacketAsync(msg, 10);
-
-                await Task.Delay(Constants.DELAY_BETWEEN_MESSAGES);
-                await AssertUtils.ContainsWithRetriesAsync("+MSG: Done", ArduinoDevice.SerialLogs);
-                TestFixture.ClearLogs();
-            }
-
-            var c2dMessage = new LoRaCloudToDeviceMessage()
-            {
-                Fport = 0,
-                Payload = string.Empty,
-                MacCommands = { new LinkADRRequest(3, 4, 25, 0, 1) } // Update data rate to DR3
-            };
-
-            await TestFixtureCi.SendCloudToDeviceMessageAsync(device.DeviceID, c2dMessage);
-            Log($"C2D Message sent to device, need to check if it receives");
-
-            var foundC2DMessageCount = 0;
-            var foundLinkADRCmdCount = 0;
-            var foundChangedDataRateCount = 0;
-
-            var c2dLogMessage = $"{device.DeviceID}: cloud to device MAC command LinkADRCmd received Type: LinkADRCmd Answer, datarate: 3";
-            var linkADRCmdMessage = "LinkADRCmd mac command detected in upstream payload";
-            var changedDataRateMessage = $"'datr':'SF9BW125'"; // DR3
-
-            // Sends 8x confirmed messages, stopping if C2D message is found and data rate is updated
-            for (var i = warmUpMessageCount + 1; i <= messagesToSend; ++i)
-            {
-                var msg = PayloadGenerator.Next().ToString(CultureInfo.InvariantCulture);
-                Log($"{device.DeviceID}: Sending unconfirmed '{msg}' {i}/{messagesToSend}");
-                await ArduinoDevice.transferPacketAsync(msg, 10);
-
-                await Task.Delay(Constants.DELAY_BETWEEN_MESSAGES);
-
-                // After transferPacket: Expectation from serial
-                // +MSG: Done
-                await AssertUtils.ContainsWithRetriesAsync("+MSG: Done", ArduinoDevice.SerialLogs);
-
-                // check if C2D message was found
-                // 0000000000000032: cloud to device MAC command LinkADRCmd received Type: LinkADRCmd
-                var C2DSearchResults = await TestFixtureCi.SearchNetworkServerModuleAsync(
-                    (messageBody) =>
-                    {
-                        return messageBody.StartsWith(c2dLogMessage, StringComparison.OrdinalIgnoreCase);
-                    },
-                    new SearchLogOptions
-                    {
-                        Description = c2dLogMessage,
-                        MaxAttempts = 1
-                    });
-
-                // check if LinkADRCmd MAC Command was detected
-                // LinkADRCmd mac command detected in upstream payload
-                var linkADRCmdSearchResults = await TestFixtureCi.SearchNetworkServerModuleAsync(
-                    (messageBody) =>
-                    {
-                        return messageBody.Contains(linkADRCmdMessage, StringComparison.OrdinalIgnoreCase);
-                    },
-                    new SearchLogOptions
-                    {
-                        Description = linkADRCmdMessage,
-                        MaxAttempts = 1
-                    });
-
-                // check if the data rate was changed to DR3
-                // 'datr':'SF9BW125'
-                var changedDataRateSearchResults = await TestFixtureCi.SearchNetworkServerModuleAsync(
-                    (messageBody) =>
-                    {
-                        return messageBody.Contains(changedDataRateMessage, StringComparison.OrdinalIgnoreCase);
-                    },
-                    new SearchLogOptions
-                    {
-                        Description = changedDataRateMessage,
-                        MaxAttempts = 1
-                    });
-
-                if (C2DSearchResults.Found)
-                {
-                    foundC2DMessageCount++;
-                    Log($"{device.DeviceID}: Found C2D message in log (after sending {i}/{messagesToSend}) {foundC2DMessageCount} times");
-                    EnsureNotSeenTooManyTimes(foundC2DMessageCount);
-                }
-
-                if (linkADRCmdSearchResults.Found)
-                {
-                    foundLinkADRCmdCount++;
-                    Log($"{device.DeviceID}: Found LinkADRCmd in log (after sending {i}/{messagesToSend}) {foundC2DMessageCount} times");
-                    EnsureNotSeenTooManyTimes(foundLinkADRCmdCount);
-                }
-
-                if (changedDataRateSearchResults.Found)
-                {
-                    foundChangedDataRateCount++;
-                    Log($"{device.DeviceID}: Found changed data rate in log (after sending {i}/{messagesToSend}) {foundC2DMessageCount} times");
-                }
-
-                TestFixture.ClearLogs();
-                await Task.Delay(Constants.DELAY_BETWEEN_MESSAGES);
-
-                if (foundC2DMessageCount > 0 && foundLinkADRCmdCount > 0 && foundChangedDataRateCount > 0)
-                    break;
-            }
-
-            Assert.True(foundC2DMessageCount > 0, $"Could not find C2D message with MAC command LinkADRCmd in LNS log");
-            Assert.True(foundLinkADRCmdCount > 0, $"Could not find LinkADRCmd MAC Command in LNS log");
-            Assert.True(foundChangedDataRateCount > 0, $"Could not find updated data rate in LNS log");
         }
 
         /// <summary>
